requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
openapi-codec==1.3.2
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
reportlab==3.5.34
requests==2.24.0
simplejson==3.17.2
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
